+++ Tuneful/ViewModels/StatusBarItemManager.swift
@@ -27,9 +27,9 @@ class StatusBarItemManager: ObservableObject {
         }
         
         let titleWidth = title.stringWidth(with: Constants.StatusBar.marqueeFont)
-        let menuBarItemWidth = titleWidth == 0 ? Constants.StatusBar.imageWidth : (self.menuBarItemWidth > titleWidth ? titleWidth : self.menuBarItemWidth)
+        let menuBarItemWidth = titleWidth == 0 ? Constants.StatusBar.imageWidth : (self.menuBarItemWidth > titleWidth  ? titleWidth + 40 : self.menuBarItemWidth)
         let isItemBiggerThanLimit = Constants.StatusBar.imageWidth + title.stringWidth(with: Constants.StatusBar.marqueeFont) >= menuBarItemWidth
-        let xOffset = isItemBiggerThanLimit ? 10.0 : (self.menuBarItemWidth - Constants.StatusBar.imageWidth - title.stringWidth(with: Constants.StatusBar.marqueeFont)) / 2
+        let xOffset = isItemBiggerThanLimit ? 10.0 : (menuBarItemWidth - Constants.StatusBar.imageWidth - title.stringWidth(with: Constants.StatusBar.marqueeFont)) / 2
         
         let menuBarView = NSHostingView(rootView: menuBarIconView)
         menuBarView.frame = NSRect(x: xOffset, y: 1, width: menuBarItemWidth, height: 20)